Bump minor version and tag commits on push to branch
Adds a GitHub Actions workflow that bumps the minor version and commits
it with a matching tag, when an untagged commit is pushed to the branch.

Initially, the branch is set to
"increment_version_and_create_release_on_main_merge" whilst its tested.
A future commit will change it to "main".